Introduce `expected_bytes_per_token` for improved buffer size prediction.
- Standardize the use of `f32` for `EXPECTED_BYTES_PER_TOKEN`.
- Add methods for configuring and predicting buffer sizes in decoders and encoders.
- Replace hardcoded capacity calculations with reusable methods.
- Adjust related structures and documentation for consistency.

+++ crates/wordchipper/src/decoders/utility/byte_decoder.rs
@@ -43,7 +43,7 @@ impl<T: TokenType> TokenDecoder<T> for ByteDecoder<T> {
         &self,
         tokens: &[T],
     ) -> anyhow::Result<DecodeResult<Vec<u8>>> {
-        let capacity = (tokens.len() as f64 * EXPECTED_BYTES_PER_TOKEN) as usize;
+        let capacity = (tokens.len() as f32 * EXPECTED_BYTES_PER_TOKEN) as usize;
         let mut value = Vec::with_capacity(capacity);
         let mut consumed = 0;
         for &t in tokens {
