+++ crates/wordchipper/src/vocab/size_hints.rs
@@ -4,4 +4,4 @@
 pub const EXPECTED_WORD_LENGTH: usize = 5;
 
 /// Expected bytes/token ratio.
-pub const EXPECTED_BYTES_PER_TOKEN: f64 = 4.0;
+pub const EXPECTED_BYTES_PER_TOKEN: f32 = 4.0;
